# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "aurevaRewards API",
    "version": "1.0.0",
    "description": "REST API for the Aureva Rewards blockchain loyalty platform. User endpoints require a Bearer JWT from `POST /auth/login`. Merchant endpoints require an `x-api-key` header.",
    "contact": {
      "name": "Aureva Rewards Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local development (current v1)"
    },
    {
      "url": "https://api.aurevaRewards.io/api/v1",
      "description": "Production (current v1)"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Local development legacy alias for v1"
    },
    {
      "url": "https://api.aurevaRewards.io/api",
      "description": "Production legacy alias for v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /auth/login`."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Merchant API key returned on registration (shown once)."
      }
    },
...
  },
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user (wallet-based registration)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "walletAddress"
                ],
                "properties": {
                  "walletAddress": {
                    "type": "string",
                    "example": "GBBD47..."
                  },
                  "referralCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Wallet already registered"
          }
        }
      }
    },
...
    "/admin/stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Aggregate platform statistics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Platform stats"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Admin required"
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Paginated user list",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User list"
          },
          "403": {
            "description": "Admin required"
          }
        }
      }
    },
    "/admin/rewards": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all platform rewards",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rewards list"
          },
          "403": {
            "description": "Admin required"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a reward",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "pointsCost"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "pointsCost": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reward created"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Admin required"
          }
        }
      }
    },
    "/admin/rewards/{id}": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update a reward",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "pointsCost": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated reward"
          },
          "403": {
            "description": "Admin required"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a reward",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "description": "Admin required"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/admin/audit-logs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Paginated audit log",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "entityType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log"
          },
          "403": {
            "description": "Admin required"
          }
        }
      }
    },
...
  }
}